vault_client dependency replaced with ureq
