+++ Cargo.toml
@@ -62,7 +62,7 @@ yubihsm = { version = "0.41", features = ["secp256k1", "setup", "usb"], optional
 zeroize = "1"
 
 
-
+# HashiCorp deps
 ureq = { version = "~2.5", default-features=false, features = ["tls", "json", "gzip"], optional = true}
 base64 = { version = "0.13.0", optional = true}
 aes-kw = { version = "0.2.1", features = ["std"], optional = true}
