+++ src/keyring/providers/hashicorp.rs
@@ -2,7 +2,6 @@
 pub(crate) mod client;
 pub(crate) mod error;
 pub(crate) mod signer;
-pub(crate) mod vault_data;
 
 use crate::{
     chain,
